+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 1.0 (dev)
 * Allow passing ``SharedContent`` object to ``sharedcontent`` template tag.
 * Allow sharing a ``SharedContent`` object across multiple websites.
 * Improve default value of ``ContentPlugin.cache_timeout`` for Django 1.6 support.
+* Fix frontend media support for ``{% sharedcontent %}`` tag.
 
 Released on 1.0c2:
 ~~~~~~~~~~~~~~~~~~